requirements.txt
click~=8.1.6
setuptools~=65.5.0
Pillow~=10.0.0